requirements.txt
# 🔥 Core (LangGraph + LangChain stable combo)
langgraph>=0.2.0
langgraph-checkpoint-postgres>=1.0.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.3

# 🗄️ PostgreSQL (psycopg v3 required by langgraph-checkpoint-postgres)
psycopg[binary]>=3.1.0
psycopg-pool>=3.1.0

# 📚 RAG + Embeddings
chromadb==0.4.24
sentence-transformers==2.6.1
transformers==4.38.2
huggingface-hub==0.20.3

# ⚡ API
fastapi==0.110.0
uvicorn==0.29.0

# 🎨 UI
streamlit==1.32.2

# 🔧 Utils
numpy==1.26.4
tqdm==4.66.2
python-dotenv==1.0.1

# ⚠️ Compatibility
pydantic>=2.0.0